Challenges/Hard/TheKeytoAccess/app/requirements.txt
Flask==2.0.3
Flask-Session==0.4.0
Flask-Cors==3.0.1
Werkzeug==2.0.3
PyJWT==2.8.0